Add possibility of adding a suffix to generated Python for LLDB
In the end the libpython3*.dll are in the bin folder where the compiler
also lives. When adding it to the path or to an SDK "Sysroot" together
with other DLLs and binaries other Software might pick up these DLLs which
they shouldn't as this is not a regular Windows-style Python build.